fix: check if positions samples elements size is 1

+++ tasks/RudderActuatorPositionToAnalogTask.hpp
@@ -11,9 +11,9 @@ namespace gazebo_usv {
         friend class RudderActuatorPositionToAnalogTaskBase;
 
     protected:
-        float m_analog_min;
-        float m_position_min;
-        float m_analog_per_length_ratio;
+        double m_analog_min;
+        double m_position_min;
+        double m_analog_per_length_ratio;
 
         base::samples::Joints m_position;
         std::vector<raw_io::Analog> m_analog_out;
